feat: [Comment] User 객체 추가, LocalDateTime->LocalDate 수정

+++ backend/src/main/java/com/project/backend/comment/service/CommentService.java
@@ -6,6 +6,8 @@ import com.project.backend.comment.dto.CommentListResponseDto;
 import com.project.backend.comment.dto.CommentSaveRequestDto;
 import com.project.backend.post.domain.entity.Post;
 import com.project.backend.post.domain.repository.PostRepository;
+import com.project.backend.security.domain.entity.User;
+import com.project.backend.security.security.CurrentUser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -22,10 +24,10 @@ public class CommentService {
     private final CommentRepository commentRepository;
 
     @Transactional
-    public Comment save(Long pid, CommentSaveRequestDto requestDto) {
+    public Long save(Long pid, CommentSaveRequestDto requestDto, User user) {
         Optional<Post> postItem = postRepository.findById(pid);
         requestDto.setPost(postItem.get());
-        return commentRepository.save(requestDto.toCommentEntity());
+        return commentRepository.save(requestDto.toCommentEntity(user)).getCid();
     }
 
     @Transactional
@@ -37,10 +39,17 @@ public class CommentService {
     }
 
     @Transactional
-    public void delete(Long pid, Long cid){
+    public String delete(Long pid, Long cid, User user){
         Comment comment = commentRepository.findById(cid)
         .orElseThrow(()->new
                 IllegalArgumentException("해당 댓글이 없습니다."));
-        commentRepository.delete(comment);
+        if((comment.getUser()).getUid().equals(user.getUid())){
+            commentRepository.delete(comment);
+            return "댓글이 삭제되었습니다.";
+        }
+        else{
+            return "해당 글쓴이가 아닙니다.";
+        }
+
     }
 }
